fixed teleport not removing during onmount
